repository/src/androidTest/java/ru/dopaminka/persistence/ProgramProviderImplTest.kt
package ru.dopaminka.persistence

import android.content.res.Resources
import androidx.annotation.RawRes
import androidx.test.ext.junit.runners.AndroidJUnit4
import androidx.test.platform.app.InstrumentationRegistry
import org.junit.Assert.assertTrue
import org.junit.Test
import org.junit.runner.RunWith

/**
 * Instrumented test, which will execute on an Android device.
 *
 * See [testing documentation](http://d.android.com/tools/testing).
 */
@RunWith(AndroidJUnit4::class)
class ProgramProviderImplTest {
    @Test
    fun useAppContext() {
        // Context of the app under test.
        val appContext = InstrumentationRegistry.getInstrumentation().targetContext

        val loader = ProgramProviderImpl(object : RawFileTextProvider {
            override fun get(id: Int) = appContext.resources.getRawTextFile(id)
        })
        val program = loader.get()

        assertTrue(program.lessons.isNotEmpty())
    }
}

fun Resources.getRawTextFile(@RawRes id: Int) =
    openRawResource(id).bufferedReader().use { it.readText() }